chore(refactor): refactor functions methods middlewares
* created JWT middleware for token auth
* create isAdmin middleware
* change get token in before all tests
