+++ server/test/users.test.js
@@ -13,10 +13,14 @@ describe('Users Tests', () => {
   before((done) => {
     chai
       .request(app)
-      .get('/get-token')
+      .post('/api/v1/auth/signin')
+      .send({
+        email: 'nignanthomas@gmail.com',
+        password: 'qwerty',
+      })
       .end((err, res) => {
         const result = JSON.parse(res.text);
-        token = result.token;
+        token = result.data.token;
         done();
       });
   });
