+++ server/v1/middlewares/isAdmin.js
@@ -1,0 +1,12 @@
+import { responseError } from '../helpers/responseHelpers';
+
+const isAdmin = (req, res, next) => {
+  if (!req.user.is_admin) {
+    return responseError(res, 400, 'Unauthorized! Only admin can create a trip!');
+  }
+  console.log('passed middleware');
+  return next();
+};
+module.exports = {
+  isAdmin,
+};
