+++ server/v1/routes/index.js
@@ -1,23 +1,13 @@
 import express from 'express';
-import passport from 'passport';
 import Trip from '../controllers/trips';
 import User from '../controllers/users';
 import Booking from '../controllers/bookings';
 import SignUp from '../controllers/auth/signup';
 import SignIn from '../controllers/auth/signin';
-// eslint-disable-next-line no-unused-vars
-import passportConf from '../passport';
+import { passportJWT } from '../middlewares/passportJWT';
+import { isAdmin } from '../middlewares/isAdmin';
 
 const router = express.Router();
-const passportJWT = (req, res, next) => {
-  // eslint-disable-next-line consistent-return
-  passport.authenticate('jwt', { session: false }, (err, user, info) => {
-    if (err) return next(err);
-    if (!user) return res.status(401).json({ status: 'error', error: 'Missing or Invalid Token!' });
-    req.user = user;
-    next();
-  })(req, res, next);
-};
 
 
 // SignUp route
@@ -26,16 +16,16 @@ router.post('/auth/signup', SignUp.signUp);
 router.post('/auth/signin', SignIn.signIn);
 
 // User routes
-router.get('/users', passportJWT, User.getAllUsers);
-router.get('/users/:userId', passportJWT, User.getOneUser);
+router.get('/users', passportJWT, isAdmin, User.getAllUsers);
+router.get('/users/:userId', passportJWT, isAdmin, User.getOneUser);
 
 // Trips routes
-router.post('/trips', passportJWT, Trip.createTrip);
+router.post('/trips', passportJWT, isAdmin, Trip.createTrip);
 router.get('/trips', passportJWT, Trip.getAllTrips);
 router.get('/trips/:tripId', passportJWT, Trip.getOneTrip);
-router.patch('/trips/:tripId', passportJWT, Trip.updateTrip);
-router.patch('/trips/:tripId/cancel', passportJWT, Trip.cancelTrip);
-router.delete('/trips/:tripId', passportJWT, Trip.deleteTrip);
+router.patch('/trips/:tripId', passportJWT, isAdmin, Trip.updateTrip);
+router.patch('/trips/:tripId/cancel', passportJWT, isAdmin, Trip.cancelTrip);
+router.delete('/trips/:tripId', passportJWT, isAdmin, Trip.deleteTrip);
 
 // Bookings routes
 router.post('/bookings', passportJWT, Booking.createBooking);
